Extract yearly reading goal code
