+++ openlibrary/plugins/upstream/code.py
@@ -31,6 +31,7 @@ from openlibrary.plugins.upstream import (
     edits,
     merge_authors,
     models,
+    readinggoals,
     recentchanges,  # noqa: F401 side effects may be needed
     spamcheck,
     utils,
@@ -383,6 +384,7 @@ def setup():
     # merge_works.setup() # ILE code
     edits.setup()
     checkins.setup()
+    readinggoals.setup()
 
     from openlibrary.plugins.upstream import data, jsdef
