+++ openlibrary/templates/account/mybooks.html
@@ -99,7 +99,7 @@ $def empty_mobile_carousel(data):
             </div>
             $if current_goal:
               <span id="reading-goal-container">
-                $:render_template('check_ins/reading_goal_progress', [current_goal])
+                $:render_template('reading_goals/reading_goal_progress', [current_goal])
               </span>
           </div>
         </li>
