+++ static/css/components/work.less
@@ -315,5 +315,6 @@ div.editionTools {
 }
 
 @import (less) "components/check-in.less";
+@import (less) "components/reading-goal.less";
 @import (less) "components/loading-indicator.less";
 @import (less) "components/work-title-and-author.less";
